requirements.txt

llama-cpp-python
huggingface-hub





